requirements.txt
discord.py
praw
requests